(PC-10080) Add Hero ui component
